mode ProductOptions as POCO and moved the methods to repo
