+++ refactor-me/Models/ProductOptions.cs
@@ -1,38 +1,14 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Data.SqlClient;
+﻿using System.Collections.Generic;
 
 namespace refactor_me.Models
 {
     public class ProductOptions
     {
         public List<ProductOption> Items { get; private set; }
-        private readonly ProductOptionRepository _productOptionRepository = new ProductOptionRepository();
 
         public ProductOptions()
         {
-            LoadProductOptions(null);
-        }
-
-        public ProductOptions(Guid productId)
-        {
-            LoadProductOptions($"where productid = '{productId}'");
-        }
-
-        private void LoadProductOptions(string where)
-        {
             Items = new List<ProductOption>();
-            var conn = Helpers.NewConnection();
-            var cmd = new SqlCommand($"select id from productoption {where}", conn);
-            conn.Open();
-
-            var rdr = cmd.ExecuteReader();
-            while (rdr.Read())
-            {
-                var id = Guid.Parse(rdr["id"].ToString());
-                var option = _productOptionRepository.Get(id);
-                Items.Add(option);
-            }
         }
     }
 }
